+++ include/GafferML/ImageToTensor.h
@@ -69,6 +69,9 @@ class GAFFERML_API ImageToTensor : public Gaffer::ComputeNode
 		Gaffer::BoolPlug *interleaveChannelsPlug();
 		const Gaffer::BoolPlug *interleaveChannelsPlug() const;
 
+		Gaffer::StringPlug *tensorDataTypePlug();
+		const Gaffer::StringPlug *tensorDataTypePlug() const;
+
 		TensorPlug *tensorPlug();
 		const TensorPlug *tensorPlug() const;